Relicense project to LGPL-3.0
The previous decision to switch to GPL-3.0 was based on pycddlib being a
dependency, since pycddlib is under the GPL-2.0. But pycddlib is not a
required dependency of lpsolvers any more.

When a solver (e.g. pycddlib or CVXOPT, both GPL) is available,
lpsolvers will provide an interface to it, but lpsolvers itself does not
depend on any particular solver. That is, it is separate from individual
solvers.